WIP: Merge duplicate funcs 3
